+++ examples/parse_csv.cpp
@@ -12,10 +12,8 @@
 #include <ctb/tables/WineListTable.h>
 
 #include <cassert>
-#include <deque>
 #include <print>
 #include <string>
-#include <vector>
 
 #pragma warning(push)
 #pragma warning(disable: 4365 4464 4702)
@@ -31,8 +29,8 @@ int main()
    try
    {
       csv::CSVReader reader{ R"(C:\Users\jkohn\AppData\Roaming\cts_win\List.csv)" };
-      std::deque<WineListRecord> wines{};
-      WineListRecord rec{};
+      WineListTable wines{};
+      WineListTable::value_type rec{};
       for (csv::CSVRow& row : reader)
       {
          rec.parse(row);
