+++ lib/CMakeLists.txt
@@ -15,6 +15,8 @@ target_sources(ctBrowse_lib
       "../include/ctb/CredentialWrapper.h"
       "../include/ctb/Error.h"
       "../include/ctb/log.h"
+      "../include/ctb/table_data.h"
+      "../include/ctb/table_download.h"
       "../include/ctb/utility.h"
       "../include/ctb/utility_chrono.h"
       "../include/ctb/utility_http.h"
@@ -32,15 +34,14 @@ target_sources(ctBrowse_lib
       "../include/ctb/tables/CtDataTable.h"
       "../include/ctb/tables/PendingWineTable.h"
       "../include/ctb/tables/WineListTable.h"
-      "../include/ctb/tables/PropFilter.h"
-      "../include/ctb/tables/PropStringFilter.h"
-      "../include/ctb/tables/PropStringFilterMgr.h"
-      "../include/ctb/tables/SubstringFilter.h"
-      "../include/ctb/tables/TableProperty.h"
-      "../include/ctb/tables/TableRecord.h"
-      "../include/ctb/tables/TableSorter.h"
-      "../include/ctb/tables/table_data.h"
-      "../include/ctb/tables/table_download.h"
+
+      "../include/ctb/tables/detail/PropFilter.h"
+      "../include/ctb/tables/detail/PropStringFilter.h"
+      "../include/ctb/tables/detail/PropStringFilterMgr.h"
+      "../include/ctb/tables/detail/SubstringFilter.h"
+      "../include/ctb/tables/detail/TableProperty.h"
+      "../include/ctb/tables/detail/TableRecord.h"
+      "../include/ctb/tables/detail/TableSorter.h"
 
       "../include/ctb/tasks/PollingTask.h"
       "../include/ctb/tasks/tasks.h"
