+++ lib/table_download.cpp
@@ -5,7 +5,7 @@
  *
  * @copyright  Copyright © 2025 Jeff Kohn. All rights reserved.
  *********************************************************************/
-#include "ctb/tables/table_download.h"
+#include "ctb/table_download.h"
 #include "ctb/utility.h"
 #include "ctb/utility_http.h"
 #include "external/HttpStatusCodes.h"
